backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
selenium==4.29.0
python-dotenv==1.0.1
undetected-chromedriver==3.5.5
asyncio==3.4.3
pydantic==2.10.6
